Moving key-state.js to utils folder
