+++ src/module/demonlord.js
@@ -6,7 +6,7 @@ import {DemonlordItem} from './item/item.js'
 import {ActionTemplate} from './pixi/action-template.js'
 import {registerSettings} from './settings.js'
 import {registerVisionModes} from './vision.js'
-import KeyState from './key-state.js'
+import KeyState from './utils/key-state.js'
 import {DLCombatTracker} from './combat/combat-tracker.js'
 import {preloadHandlebarsTemplates} from './templates.js'
 import * as migrations from './migration.js'
